requirements-prod.txt
# Production API + hybrid inference (Docker / Render).
# For notebooks, training, ablations, and full EDA: use requirements.txt

# --- Core numerics & ML ---
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# --- Deep learning & NLP ---
torch>=2.0.0
transformers>=4.40.0
sentencepiece>=0.1.99
accelerate>=0.27.0
peft>=0.10.0

# --- Graph (hybrid_pipeline still loads BN modules when imported) ---
networkx==3.2.1
pgmpy==0.1.25

# --- Web API ---
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# --- Documents & OCR ---
# Headless OpenCV only (avoids GUI stack); EasyOCR-compatible
opencv-python-headless>=4.9.0.80
pdfplumber==0.10.3
Pillow==10.2.0
pdf2image==1.17.0
easyocr==1.7.1

# --- Phase 2 configs ---
PyYAML>=6.0.0
